refactor: remove namespace of view
